punctuation update to local-git-setup.md
Request Review. I have made one additional change to the local-git-setup.md file at Line 120, adding a semi-colon ":" to the end of 
"Next head over to GitHub, and log into your account" so that it displays as "Next head over to GitHub, and log into your account:". I 
am suggesting this change as a bulleted list is being used here. Best practices indicate a colon should be used to introduce the list.