Update git-in-here.yml
Thank you for your contributions, your projects are very cool, a lot of difficulty in learning these new technologies alone, the courses are very expensive in BR. Their designs are beautiful and very creative. Your advice to start with a project that we are passionate about is perfect, I work mainly as a volunteer and taking these technologies to friends who know very little and don't even know how to use a computer, there are still people like that... I love your work, soon I will be sponsoring. 🫶